1.17.1

+++ net/minecraft/client/realms/task/DownloadTask.java
@@ -34,7 +34,7 @@ public class DownloadTask extends LongRunningTask {
 				}
 
 				WorldDownload worldDownload = realmsClient.download(this.worldId, this.slot);
-				pause(1);
+				pause(1L);
 				if (this.aborted()) {
 					return;
 				}
@@ -47,7 +47,7 @@ public class DownloadTask extends LongRunningTask {
 					return;
 				}
 
-				pause(var4.delaySeconds);
+				pause((long)var4.delaySeconds);
 				i++;
 			} catch (RealmsServiceException var5) {
 				if (this.aborted()) {
